Extract HandleTagNameMixin from mixin.rb
